pinecone/pyproject.toml
[tool.poetry]
name = "pinecone"
version = "0.1.0"
description = ""
authors = ["StarmiyaMiyuki <sousaku_0912@yahoo.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pinecone-client = "^4.1.0"
python-dotenv = "^1.0.1"
openai = "^1.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
